+++ package/thunderstore/MyLittleUI/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.13
+* new config to disable minimap in normal mode and adapt minimap blocks for nomap layout
+* new config to disable mass repair
+
 # 1.2.12
 * minor technical optimizations